+++ BNA_KC_Weapons/Config.cpp
@@ -326,198 +326,6 @@ class CfgWeapons
 		};
 	};
 
-    // Basically a fake UGL to get around the camera and cursor being odd on pistols
-    // Values come from UGL_F base class
-    class BNA_KC_Launcher_Muzzle: Rifle_Base_F
-    {
-        displayName = "Launcher";
-        magazines[] = {};
-        magazineWell[] = {};
-
-        modes[] = { "Single" };
-        muzzles[] = { "this" };
-
-        nameSound = "";
-
-        changeFiremodeSound[] = { "A3\Sounds_F\arsenal\weapons\UGL\Firemode_ugl", 0.316228, 1, 5 };
-        reloadMagazineSound[] = { "A3\Sounds_F\arsenal\weapons\UGL\Reload_UGL", 0.562341, 1, 10 };
-        drySound[] = { "A3\Sounds_F\arsenal\weapons\UGL\Dry_ugl", 0.562341, 1, 10 };
-
-        recoil = "recoil_pistol_p07";
-        recoilProne = "recoil_pistol_p07";
-
-        cursor = "EmptyCursor";
-        cursorAim = "gl";
-        textureType = "semi";
-
-        autoFire = 0;
-
-        class Single: Mode_SemiAuto
-        {
-            textureType = "semi";
-
-            // recoil = "recoil_pistol_p07";
-            // recoilProne = "recoil_pistol_p07";
-
-            // For AI units
-            maxRange = 400;
-            maxRangeProbab = 0.05;
-            midRange = 200;
-            midRangeProbab = 0.7;
-            minRange = 30;
-            minRangeProbab = 0.1;
-
-            nameSound = "";
-            sounds[] = { "StandardSound" };
-            class BaseSoundModeType
-            {
-                closure1[] = { "A3\Sounds_F\arsenal\weapons\UGL\Closure_UGL", 1, 1, 10 };
-                soundClosure[] = { "closure1", 1 };
-            };
-            class StandardSound: BaseSoundModeType
-            {
-                begin1[] = { "A3\Sounds_F\arsenal\weapons\UGL\UGL_01", 0.707946, 1, 200 };
-                begin2[] = { "A3\Sounds_F\arsenal\weapons\UGL\UGL_02", 0.707946, 1, 200 };
-                soundBegin[] = { "begin1", 0.5, "begin2", 0.5 };
-
-                soundSetShot[] = { "UGL_shot_SoundSet", "UGL_Tail_SoundSet", "UGL_InteriorTail_SoundSet" };
-            };
-        };
-    };
-
-    class hgun_P07_F;
-    class JLTS_DC17SA: hgun_P07_F
-    {
-        class Single: Mode_SemiAuto
-        {
-            class BaseSoundModeType {};
-            class StandardSound: BaseSoundModeType {};
-        };
-
-        class Stun;
-
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            class CowsSlot;
-        };
-    };
-    class BNA_KC_DC17: JLTS_DC17SA
-    {
-        // Mod Info
-        dlc = "BNA_KC";
-        author = "SweMonkey and DartRuffian";
-
-        // Scope
-        scope = 2;
-        scopeArsenal = 2;
-
-        displayName = "[KC] DC-17";
-        baseWeapon = "BNA_KC_DC17"; // Used to specify it as a different weapon
-        canShootInWater = 1;
-
-        JLTS_shieldedWeapon = "BNA_KC_DC17_RiotShield";
-
-        magazines[] = { "12thFleet_Mag_DC17" };
-        modes[] = { "Single" };
-        muzzles[] = { "this", "Stun", "Launcher" };
-
-        class Single: Single
-        {
-            class StandardSound: StandardSound
-            {
-                begin1[] = {"BNA_KC_Gear\Weapons\Data\Audio\BNA_KC_DC17_Fire1", 1, 1, 1800};
-                soundBegin[] = {begin1, 1};
-            };
-        };
-
-        class Launcher: BNA_KC_Launcher_Muzzle
-        {
-            displayName = "Flare Launcher";
-            magazines[] =
-            {
-                "3Rnd_UGL_FlareWhite_F",
-                "3Rnd_UGL_FlareRed_F",
-                "BNA_KC_3Rnd_UGL_FlareBlue"
-            };
-        };
-
-        class Stun: Stun
-        {
-            displayName = "Stun";
-            magazines[] =
-            {
-                "12thFleet_Mag_StunLong",
-                "12thFleet_Mag_StunShort"
-            };
-            magazineWell[] = {};
-        };
-
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            class CowsSlot
-            {
-                compatibleItems[] = 
-                {
-                    // Sights
-                    "aux501_cows_pistol",
-                    "aux501_cows_pistol_2"
-                };
-                
-                displayName = "Optics Slot";
-                iconPicture = "\A3\Weapons_F\Data\UI\attachment_top.paa";
-                iconPinpoint = "Bottom";
-                iconPosition[] = { 0.5, 0.35 };
-                iconScale = 0.2;
-                linkProxy = "\A3\data_f\proxies\weapon_slots\TOP";
-                scope = 0;
-            };
-        };
-    };
-
-    class BNA_KC_DC17_RiotShield: BNA_KC_DC17
-    {
-        // Scope
-        scope = 1;
-        scopeArsenal = 0;
-
-        displayName = "[KC] DC-17 (Shield)";
-
-        model = "\MRC\JLTS\weapons\DC17SA\DC17SA_shielded.p3d";
-        handAnim[] = {"OFP2_ManSkeleton","\MRC\JLTS\weapons\DC17SA\anims\DC17SA_shielded_handanim.rtm"};
-
-        JLTS_baseWeapon = "BNA_KC_DC17";
-        JLTS_isShielded = 1;
-
-        recoil = "recoil_pdw"; // More recoil since the gun is held in one hand
-
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            class UnderBarrelSlot: UnderBarrelSlot
-            {
-                compatibleItems[] = {"JLTS_riot_shield_attachment"};
-            };
-        };
-    };
-
-    class BNA_KC_GR20: BNA_KC_DC17
-    {
-        displayName = "[KC] GR-20";
-        descriptionShort = "An old GR-10 plasma blaster, modified to shoot <br />pressurized bacta canisters instead of plasma.";
-        baseWeapon = "BNA_KC_GR20"; // Used to specify it as a different weapon
-
-        magazines[] = {};
-        magazineWell[] = {};
-
-        modes[] = {};
-        muzzles[] = { "Launcher" };
-
-        class Launcher: Launcher
-        {
-            displayName = "Bacta Launcher";
-            magazines[] = { "BNA_KC_Mag_GR20" };
-        };
-    };
-
     // ┌────────────────────┐
     // │      Grenades      │
     // └────────────────────┘
@@ -637,32 +445,6 @@ class CfgMagazines
         BNA_KC_GrenadeBacta_Radius = 20;
         BNA_KC_GrenadeBacta_Duration = 20;
     };
-
-    class 1Rnd_SmokeBlue_Grenade_shell;
-    class BNA_KC_Mag_GR20: 1Rnd_SmokeBlue_Grenade_shell
-    {
-        // Non-grenade version for bacta launcher
-        // Mod Info
-        dlc = "BNA_KC";
-        author = "SweMonkey and DartRuffian";
-
-        // Scope
-        scope = 2;
-        scopeArsenal = 2;
-
-        displayName = "[KC] Pressurized Bacta Canister";
-        displayNameShort = "Bacta Bomb";
-        descriptionShort = "Pressurized bacta gas. <br />Nicknamed the 'Bacta Bomb'.";
-
-        picture = "\BNA_KC_Gear\Weapons\Data\Textures\UI\BNA_KC_Magazine_BactaBomb_UI.paa";
-
-        ammo = "BNA_KC_Grenade_BactaBomb_Ammo";
-        count = 1;
-
-        BNA_KC_GrenadeType = "BACTA";
-        BNA_KC_GrenadeBacta_Radius = 10;
-        BNA_KC_GrenadeBacta_Duration = 15;
-    };
 };
 
 class CfgAmmo
@@ -780,41 +562,3 @@ class CfgAmmo
 
 
 // Particle Effects Group, used for smoke grenades
-class SmokeShellWhiteEffect;
-class BNA_KC_EffectsGroup_BactaSmoke: SmokeShellWhiteEffect
-{
-    // Set of particle effects, multiple effects for variance
-    class Smoke
-    {
-        type = "BNA_KC_Effects_BactaSmoke"; // CfgCloudlets class
-        simulation = "particles";
-
-        intensity = 1;
-        interval = 1;
-        position[] = {0, 0, 0};
-    };
-    class SmokeUW: Smoke
-    {
-        type = "BNA_KC_Effects_BactaSmokeUW";
-    };
-};
-
-
-class CfgCloudlets
-{
-    class SmokeShellWhite;
-    class BNA_KC_Effects_BactaSmoke: SmokeShellWhite
-    {
-        interval = 0.01;
-        ignoreWind = true;
-    };
-
-    // Underwater effects, easier to just inherit from the base underwater
-    // particles than to change the same properties they do
-    class SmokeShellWhiteUW;
-    class BNA_KC_Effects_BactaSmokeUW: SmokeShellWhiteUW
-    {
-        interval = 0.01;
-        ignoreWind = true;
-    };
-};
